feat(QuanTrong): Đây là commit quan trọng nhất, quyết định sự sống còn của cả hệ thống
- Thuc hien xoa benh nhan, nhung de xoa duoc benh nhan phai xoa toan bo cac bang lien qua va cac bang con cua no.
- Quy trinh xoa: benh nhan goi phuong thuc Detete cua HoSoYTeService -> ket qua dieu tri -> don thuoc -> trieu chung -> ket qua xet nghiem -> ho so y te.

+++ Server/HeThongHoTroVaQuanLyPhongKham/Services/DonThuocService.cs
@@ -50,11 +50,13 @@ namespace HeThongHoTroVaQuanLydonThuoc.Services
 
         public async Task DeleteAsync(int id)
         {
-            await _ketQuaDieuTriService.DeleteByMaHoSoYTeAsync(id);
+            var donThuoc = await GetByIdAsync(id);
+           
+            await _ketQuaDieuTriService.DeleteByMaHoSoYTeAsync(donThuoc.MaHoSoYte);
 
             await _donThuocRepository.DeleteAsync(
                 _donThuocMapping.MapDtoToEntity(
-                    await GetByIdAsync(id)));
+                    donThuoc));
         }
 
         public async Task DeleteByMaHoSoYTeAsync(int id)
@@ -62,7 +64,8 @@ namespace HeThongHoTroVaQuanLydonThuoc.Services
             var donthuocs = await _donThuocRepository.GetQueryable()
                                 .Where(dt => dt.MaHoSoYte == id)
                                 .ToListAsync();
-            await _donThuocRepository.DeleteAsync(donthuocs);
+            if (donthuocs.Any())
+                await _donThuocRepository.DeleteAsync(donthuocs);
         }
 
         public async Task<IEnumerable<DonThuocDTO>> GetAllAsync(int page, int pageSize)
